requirements.txt
numpy
python-dotenv
torch
sentence-transformers
scikit-learn
spacy
fastapi
uvicorn[standard]
motor
redis
pydantic
pydantic-settings
google-generativeai
pyyaml
langchain
langchain-community
langchain-qdrant
qdrant-client
agno
rank_bm25
pytest 
pytest_asyncio
httpx
prometheus-fastapi-instrumentator
fuzzywuzzy
python-Levenshtein
passlib[bcrypt]
python-jose[cryptography] 
fastapi[all]

